# new-server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BKRM Retail Chain Management",
        "description": "L5 Swagger API for BKRM Retail Chain Management",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/api/supplier": {
            "get": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Get all suppliers",
                "operationId": "getSuppliers",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_type",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Supplier"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Create a new supplier",
                "operationId": "createSupplier",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSupplierInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Supplier"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/supplier/{supplier_id}": {
            "get": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Get a supplier",
                "operationId": "getSupplier",
                "parameters": [
                    {
                        "name": "supplier_id",
                        "in": "path",
                        "description": "Supplier ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Supplier"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Update a supplier",
                "operationId": "updateSupplier",
                "parameters": [
                    {
                        "name": "supplier_id",
                        "in": "path",
                        "description": "Supplier ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSupplierInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Supplier"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Delete a supplier",
                "operationId": "deleteSupplier",
                "parameters": [
                    {
                        "name": "supplier_id",
                        "in": "path",
                        "description": "Supplier ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Supplier"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/work-schedule": {
            "post": {
                "tags": [
                    "Work Schedule"
                ],
                "summary": "Create a work schedule",
                "operationId": "createWorkSchedule",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateWorkScheduleInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "message"
                                    ],
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "UpsertTime": {
                "required": [
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Supplier": {
                "required": [
                    "id",
                    "name",
                    "code"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "tax_number": {
                        "type": "string"
                    },
                    "note": {
                        "type": "string"
                    },
                    "creator_id": {
                        "type": "integer"
                    }
                },
                "type": "object",
                "oneOf": [
                    {
                        "$ref": "#/components/schemas/UpsertTime"
                    }
                ]
            },
            "CreateSupplierInput": {
                "required": [
                    "name",
                    "phone"
                ],
                "type": "object",
                "oneOf": [
                    {
                        "$ref": "#/components/schemas/UpdateSupplierInput"
                    }
                ]
            },
            "UpdateSupplierInput": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "tax_number": {
                        "type": "string"
                    },
                    "note": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "CreateWorkScheduleInput": {
                "required": [
                    "shift_id",
                    "employee_ids",
                    "date"
                ],
                "properties": {
                    "shift_id": {
                        "type": "integer"
                    },
                    "employee_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    },
                    "date": {
                        "type": "string"
                    }
                },
                "type": "object",
                "oneOf": [
                    {
                        "$ref": "#/components/schemas/UpdateWorkScheduleInput"
                    }
                ]
            },
            "UpdateWorkScheduleInput": {
                "properties": {
                    "note": {
                        "type": "string"
                    },
                    "is_absent": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            }
        }
    },
...
}